refactor: move AppStackNavigator files from folder to the top level
