+++ src/navigators/appStackNavigator.settings.ts
